Feat: create personalized error class

+++ src/storage/group/groupCreate.ts
@@ -1,10 +1,15 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "@storage/storageConfig";
+import { AppError } from "@utils/AppError";
 import { groupsGetAll } from "./groupsGetAll";
 
 export async function groupCreate(newGroup: string) {
   try {
     const storagedGroups = await groupsGetAll();
+    const groupAlreadyExists = storagedGroups.includes(newGroup);
+    if (groupAlreadyExists) {
+      throw new AppError("Já existe um grupo cadastrado com esse nome.");
+    }
     const storage = JSON.stringify([...storagedGroups, newGroup]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
   } catch (error) {
